refactor(exception): static error class & remove exception handler
