+++ src/exception/index.ts
@@ -1,7 +1,1 @@
-import ExceptionHandler from './handler';
-
-export * from './error';
-
-export {
-  ExceptionHandler,
-};
+export * from './impl';
